Implemented join URLs, moved joinchannel functionality to utility method, added BOT_NAME environment variable
