+++ src/environment.d.ts
@@ -4,6 +4,7 @@ declare global {
             NODE_ENV: 'dev' | 'prod';
             BOT_TOKEN: string;
             OWNER_ID: number;
+            BOT_NAME: string;
         }
     }
 }
